+++ lib/components/sticky_header/gf_sticky_header_builder.dart
@@ -2,11 +2,12 @@ import 'package:flutter/material.dart';
 import 'package:getwidget/getwidget.dart';
 
 typedef StickyHeaderWidgetBuilder = Widget Function(
-    BuildContext context, double stuckValue);
-
-/// Place this widget inside a [ListView], [GridView], [CustomScrollView], [SingleChildScrollView] or similar.
+  BuildContext context,
+  double stuckValue,
+);
 
 class GFStickyHeaderBuilder extends StatefulWidget {
+  /// Place this widget inside a [ListView], [GridView], [CustomScrollView], [SingleChildScrollView] or similar.
   /// Constructs a new [GFStickyHeaderBuilder] widget.
   const GFStickyHeaderBuilder({
     Key key,